Extracted clickListener Interface and fixed linting
