+++ .gitignore
@@ -72,3 +72,4 @@ lint/outputs/
 lint/tmp/
 # lint/reports/
 
+.DS_Store
