+++ app/src/main/java/com/example/fakenews/utils/clickListener.kt
@@ -1,0 +1,43 @@
+package com.example.fakenews.utils
+
+import android.content.Context
+import android.view.GestureDetector
+import android.view.MotionEvent
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+
+interface ClickListener {
+    fun onClick(view: View, position: Int)
+}
+
+internal class RecyclerTouchListener(
+    context: Context,
+    recyclerView: RecyclerView,
+    private val clickListener: ClickListener?)
+    : RecyclerView.OnItemTouchListener {
+    private var gestureDetector: GestureDetector? = null
+
+    init {
+        gestureDetector = GestureDetector(context, object: GestureDetector.SimpleOnGestureListener(){
+            override fun onSingleTapUp(e: MotionEvent?): Boolean {
+                return true
+            }
+        })
+    }
+
+    override fun onInterceptTouchEvent(rv: RecyclerView, e: MotionEvent): Boolean {
+        val child = rv.findChildViewUnder(e.x, e.y)
+        if (child != null && clickListener != null && gestureDetector!!.onTouchEvent(e)){
+            clickListener.onClick(child, rv.getChildPosition(child))
+        }
+        return false
+    }
+
+    override fun onTouchEvent(rv: RecyclerView, e: MotionEvent) {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
+
+    override fun onRequestDisallowInterceptTouchEvent(disallowIntercept: Boolean) {
+        TODO("not implemented") //To change body of created functions use File | Settings | File Templates.
+    }
+}
